program/rust: Instruction docs: Add the permissions account  (#426)
* program/rust: Consistent instruction account doc style

Indentation of the account description was inconsistent in a few cases.
Plus all instructions use `account[i]` prefix for the account
description, except for `DelProduct` and `UpdPermissions`.  That also
seems inconsistent.

* program/rust: Mention permissions account in instructions

Many instructions require the permissions account, but do not document
that it needs to be present in the instruction documentation.